==> kuro-beer/notebook/00_convert_rawdata_to_feather.ipynb <==
import gc
import time
import numpy as np
import pandas as pd

ls -lah ../data/

path = '../data/'
dtypes = {
        'ip'            : 'uint32',
        'app'           : 'uint16',
        'device'        : 'uint16',
        'os'            : 'uint16',
        'channel'       : 'uint16',
        'is_attributed' : 'uint8',
        'click_id'      : 'uint32'
        }
train_columns = ['ip', 'app', 'device', 'os', 'channel', 'click_time', 'is_attributed']
test_columns  = ['ip', 'app', 'device', 'os', 'channel', 'click_time', 'click_id']

### Read csv


%%time
train = pd.read_csv(path + 'raw/train.csv', 
                    dtype = dtypes,
                    usecols = train_columns,
                    engine = 'c')
print(train.shape)

train.dtypes

# drop messy records
train = train[train.click_time >= '2017-11-06 16:00:00'] 
print(train.shape)
train_size = train.shape[0]
gc.collect()

%%time
test = pd.read_csv(path + 'raw/test.csv', 
                   dtype = dtypes,
                   usecols = test_columns,
                   engine = 'c')
test_size = test.shape[0]
print(test.shape)

merge = pd.concat([train, test])
print(merge.shape)

merge.head()

del train, test
gc.collect()

click_time_numeric = merge.click_time.astype('datetime64[ns]')
merge = merge.drop(['click_time'], axis=1)
merge['click_time'] = click_time_numeric

print(merge.shape)
merge.head()

merge.reset_index(inplace=True)

merge.dtypes

# drop original index
merge = merge.drop(['index'], axis=1)

### fillna and create 'is_test' variable

print(np.min(merge.click_id), np.max(merge.click_id))

merge['click_id'] = merge.click_id.fillna(99999999).astype('uint32')

merge['is_attributed'] = merge.is_attributed.fillna(99).astype('uint8')

merge['is_test'] = pd.Series([0]*train_size + [1]*test_size, dtype='bool')

print(merge.shape)
merge.head()

merge.dtypes

merge.to_feather(path + 'click_data.feather')
