==> kuro-beer/notebook/01_create_unique_id_values.ipynb <==
import gc
import time
import numpy as np
import pandas as pd
from sklearn.preprocessing import LabelEncoder

ls -lah ../data/

%time
merge = pd.read_feather("../data/merge.feather", nthreads=4)

train_size = merge[merge.is_test == 0].shape[0]

merge.shape

merge.dtypes

### get time values

%%time
merge['dow'] = merge.click_time.dt.dayofweek.astype('uint8')
merge['hour'] = merge.click_time.dt.hour.astype('uint8')
merge['minute'] = merge.click_time.dt.minute.astype('uint8')
merge['second'] = merge.click_time.dt.second.astype('uint8')

click_time_5min = ((merge.click_time.dt.minute/5).apply(lambda x: x//1)*5).astype('uint8')

%%time
MOD = int(1e+8)
seconds = (merge.hour.astype('uint32')*60*60
           + merge.minute.astype('uint32')*60
           + merge.second.astype('uint32'))
merge['sin_time'] = np.sin(2*np.pi*seconds/(24*60*60)).round(8) + 0
merge['cos_time'] = np.cos(2*np.pi*seconds/(24*60*60)).round(8) + 0
del seconds

print(merge[['sin_time', 'cos_time']][merge.hour == 0].head(1))
print(merge[['sin_time', 'cos_time']][merge.hour == 6].head(1))
print(merge[['sin_time', 'cos_time']][merge.hour == 12].head(1))
print(merge[['sin_time', 'cos_time']][merge.hour == 18].head(1))

merge.dtypes

### get "ip_os_device_dow"
### get "ip_os_device_dow_hour"
### get "ip_os_device_dow_hour_5min"
ip_os_device_dowは，testデータがhour単位で虫食いなので作成しない

print('ip: ', np.min(merge.ip), np.max(merge.ip))
print('os: ', np.min(merge.os), np.max(merge.os))
print('device: ', np.min(merge.device), np.max(merge.device))

%%time
merge['ip_os_device'] = \
    (merge.device.astype('uint64')
     + merge.os.astype('uint64')*10000
     + merge.ip.astype('uint64')*10000000)

merge['ip_os_device_dow_hour'] = \
    (merge.hour.astype('uint64')
     + merge.dow.astype('uint64')*100
     + merge.device.astype('uint64')*1000
     + merge.os.astype('uint64')*10000000
     + merge.ip.astype('uint64')*10000000000)
    
merge['ip_os_device_dow_hour_5min'] = \
    (click_time_5min.astype('uint64')
     + merge.hour.astype('uint64')*100
     + merge.dow.astype('uint64')*10000
     + merge.device.astype('uint64')*100000
     + merge.os.astype('uint64')*1000000000
     + merge.ip.astype('uint64')*1000000000000)

merge.dtypes

merge.tail()

### get "ip_os_device_app_encoded"
### get "ip_os_device_channel_encoded"
stringで作ってhashすると時間かかりすぎてアレだったのでLabelEncoderを使用

print('app: ', np.min(merge.app), np.max(merge.app))
print('channel: ', np.min(merge.channel), np.max(merge.channel))

%%time
le = LabelEncoder()
merge['ip_os_device_app_encoded'] = \
    le.fit_transform(merge.ip_os_device*1000 + merge.app.astype('uint64')).astype('uint32')
del le

%%time
le = LabelEncoder()
merge['ip_os_device_channel_encoded'] = \
    le.fit_transform(merge.ip_os_device*1000 + merge.channel.astype('uint64')).astype('uint32')
del le

print(np.max(merge.ip_os_device_app_encoded))
print(np.max(merge.ip_os_device_channel_encoded))

merge.to_feather("../data/merge.feather")
