==> kuro-beer/notebook/02_down_sampling.ipynb <==
import gc
import time
from logzero import logger
import numpy as np
import pandas as pd
from sklearn.preprocessing import LabelEncoder

ls -lah ../data/

%time
merge = pd.read_feather("../data/merge.feather", nthreads=4)

train_size = merge[merge.is_test == 0].shape[0]

merge.shape

merge.dtypes

### testの時刻分布が残念なので補正

%%time
train = merge[:train_size]
test = merge[train_size:]
del merge
gc.collect()

train.tail()

test.head()

test.groupby('hour').count()

print('hour==6:  ', test[test.hour == 6].click_time.min(), test[test.hour == 6].click_time.max())
print('hour==11: ', test[test.hour == 11].click_time.min(), test[test.hour == 11].click_time.max())
print('hour==15: ', test[test.hour == 15].click_time.min(), test[test.hour == 15].click_time.max())

%%time
test['click_time'] = (test.click_time.apply(lambda x: (x - pd.offsets.Second(1)) if x.hour in [6, 11, 15] else x))

test['hour'] = test.click_time.dt.hour.astype('uint8')
test['minute'] = test.click_time.dt.minute.astype('uint8')
test['min5'] = ((test.click_time.dt.minute/5).apply(lambda x: x//1)*5).astype('uint8')
test['second'] = test.click_time.dt.second.astype('uint8')

test.groupby('hour').count()

print(test.shape)
test.dtypes

gc.collect()

### testに存在しない時間帯のデータを破棄

train = train[(train.hour >= 3) & (train.hour <= 15)]

print(train.shape)
train.groupby('hour').count()

### 統合して保存

%%time
merge = pd.concat([train, test])
del train, test
gc.collect()

%%time
merge.reset_index(inplace=True)
merge = merge.drop('index', axis=1)

merge.to_feather("../data/merge.feather")
