==> kuro-beer/notebook/10_dimension_reduction_channel.ipynb <==
import gc
import time
from logzero import logger
import numpy as np
import pandas as pd

ls -lah ../data/

%time
merge = pd.read_feather("../data/merge.feather", nthreads=8)

merge = merge[merge.is_test == False]
gc.collect()

print(merge.shape)
merge.dtypes

### preparation

%%time
channel_grp = merge.groupby('channel', as_index=False)

del merge
gc.collect()

### simple count

%%time
count_by_channel = (channel_grp.ip.count().rename(columns={'ip': 'count_by_channel'}))

count_by_channel.head()

%%time
attrib_by_channel = (channel_grp.is_attributed.sum().rename(columns={'is_attributed': 'attrib_by_channel'})).astype('uint32')

attrib_by_channel.head()

agg_by_channel = pd.concat([count_by_channel, 
                            attrib_by_channel['attrib_by_channel']], axis=1)

agg_by_channel.to_feather('../data/agg_by_channel.feather')

del agg_by_channel
gc.collect()

ls -lah ../data/*.feather

### unique count

%%time
uqcount_by_channel = pd.concat([
    channel_grp.app.nunique().to_frame('uqcount_app_by_channel')
    , channel_grp.ip.nunique().to_frame('uqcount_ip_by_channel')
    , channel_grp.device.nunique().to_frame('uqcount_device_by_channel')
    , channel_grp.os.nunique().to_frame('uqcount_os_by_channel')
    , channel_grp.ip_os_device_dow_hour.nunique().to_frame('uqcount_user_by_channel')
], axis=1)

uqcount_by_channel.head()

uqcount_by_channel.to_feather('../data/uqcount_by_channel.feather')

del uqcount_by_channel
gc.collect()

ls -lah ../data/*.feather

### cross tabulation(app)

%time
merge = pd.read_feather("../data/merge.feather", nthreads=8)

merge = merge[['channel', 'app', 'ip_os_device_dow_hour']][merge.is_test == False]
# merge = merge[['channel', 'app', 'ip_os_device_dow_hour']]
merge['counter'] = 1
gc.collect()

print(merge.shape)
merge.head()

merge.drop_duplicates(inplace = True)

print(merge.shape)
merge.head()

%%time
count_user_by_channel_app = merge.pivot_table(values  = 'counter',
                                              index   = 'channel',
                                              columns = 'app', 
                                              aggfunc = sum,
                                              fill_value = 0)

applabel = ['app_{0:03d}'.format(i) for i in list(set(merge.app))]
count_user_by_channel_app.columns = applabel
count_user_by_channel_app.reset_index(inplace=True)

count_user_by_channel_app.head()

count_user_by_channel_app.tail()

count_user_by_channel_app.to_feather('../data/count_user_by_channel_app.feather')

del count_user_by_channel_app
gc.collect()

ls -lah ../data/*.feather

### cross tabulation (channels clicked by the same user)

merge.drop(['app'], axis=1,  inplace=True)
merge.drop_duplicates(inplace=True)

print(merge.shape)
merge.head()

gc.collect()

merge['counter'] = merge['counter'].astype('bool')

d = merge.pivot('ip_os_device_dow_hour', 'channel')

d = d.fillna(False)

d.head()

%%time
mat = np.dot(d.astype('bool').astype('uint16').values.T,
             d.astype('bool').astype('uint16').values)

count_channel_coincidence_per_user = pd.DataFrame(mat)

chanlabel = ['chan_{0:03d}'.format(i) for i in list(set(merge.channel))]
count_channel_coincidence_per_user.columns = chanlabel
count_channel_coincidence_per_user.head()

del merge, mat, d
gc.collect()

count_channel_coincidence_per_user.to_feather('../data/count_channel_coincidence_per_user.feather')

del count_channel_coincidence_per_user
gc.collect()

ls -lah ../data/*.feather

### pca

import pandas as pd
import numpy as np
from sklearn.decomposition import PCA

d = pd.concat([
    pd.read_feather('../data/agg_by_channel.feather'),
    pd.read_feather('../data/uqcount_by_channel.feather'),
    pd.read_feather('../data/count_user_by_channel_app.feather'),
    pd.read_feather('../data/count_channel_coincidence_per_user.feather')
], axis = 1)

print(d.shape)
d.head()

y = d[['channel', 'attrib_by_channel']].iloc[:, 1:]
X = d.drop(['channel', 'attrib_by_channel'], axis=1)

n_components = 20
pca = PCA(n_components = n_components)
res = pca.fit_transform(X.values)

res.shape

pca.explained_variance_ratio_.cumsum()

col = ['channel_pca_{0:02d}'.format(i) for i in [i + 1 for i in range(20)]]
res = pd.DataFrame(res, columns=col)
channel_pca = pd.concat([y.channel, res], axis=1)

channel_pca.to_feather('../data/channel_pca.feather')

channel_pca.head()

### とりあえず回帰してみる

import statsmodels as sm

X1 = sm.add_constant(
    channel_pca.iloc[:, 1:] / channel_pca.iloc[:, 1:].max().max()
    , prepend=False
)

X1.head()

model = sm.GLM(y.attrib_by_channel,
               X1,
               family=sm.families.Poisson(),
               offset=X.count_by_channel)

model = sm.GLM(y.attrib_by_channel.values,
               X1.values,
               family=sm.families.Poisson(),
               offset = np.log(X.count_by_channel.values)
              )

print(y.attrib_by_channel.shape)
print(X1.shape)
print(X.count_by_channel.shape)

print(model.fit().summary())
