==> kuro-beer/notebook/11_dimension_reduction_app.ipynb <==
import gc
import time
from logzero import logger
import numpy as np
import pandas as pd

ls -lah ../data/

%time
merge = pd.read_feather("../data/merge.feather", nthreads=8)

merge = merge[merge.is_test == False]
gc.collect()

print(merge.shape)
merge.dtypes

### preparation

%%time
channel_grp = merge.groupby('app', as_index=False)

del merge
gc.collect()

### simple count

%%time
count_by_app = (channel_grp.ip.count().rename(columns={'ip': 'count_by_app'}))

print(count_by_app.shape)
count_by_app.head()

%%time
attrib_by_app = (channel_grp.is_attributed.sum().rename(columns={'is_attributed': 'attrib_by_app'})).astype('uint32')

attrib_by_app.head()

agg_by_app = pd.concat([count_by_app, 
                        attrib_by_app['attrib_by_app']], axis=1)

agg_by_app.to_feather('../data/agg_by_app.feather')

del agg_by_app
gc.collect()

ls -lah ../data/*.feather

### unique count

%%time
uqcount_by_app = pd.concat([
    channel_grp.channel.nunique().to_frame('uqcount_channel_by_app')
    , channel_grp.ip.nunique().to_frame('uqcount_ip_by_app')
    , channel_grp.device.nunique().to_frame('uqcount_device_by_app')
    , channel_grp.os.nunique().to_frame('uqcount_os_by_app')
    , channel_grp.ip_os_device_dow_hour.nunique().to_frame('uqcount_user_by_app')
], axis=1)

uqcount_by_app.head()

uqcount_by_app.to_feather('../data/uqcount_by_app.feather')

del uqcount_by_app
gc.collect()

ls -lah ../data/*.feather

### cross tabulation(channel)

%time
merge = pd.read_feather("../data/merge.feather", nthreads=8)

merge = merge[['channel', 'app', 'ip_os_device_dow_hour']][merge.is_test == False]
merge['counter'] = 1
gc.collect()

print(merge.shape)
merge.head()

merge.drop_duplicates(inplace = True)

print(merge.shape)
merge.head()

%%time
count_user_by_channel_app = merge.pivot_table(values  = 'counter',
                                              index   = 'app',
                                              columns = 'channel', 
                                              aggfunc = sum,
                                              fill_value = 0)

chlabel = ['ch_{0:03d}'.format(i) for i in list(set(merge.channel))]
count_user_by_channel_app.columns = chlabel
count_user_by_channel_app.reset_index(inplace=True)

count_user_by_channel_app.head()

count_user_by_channel_app.tail()

count_user_by_channel_app.to_feather('../data/count_user_by_app_channel.feather')

del count_user_by_channel_app
gc.collect()

ls -lah ../data/*.feather

### cross tabulation (apps clicked by the same user)

%time
merge = pd.read_feather("../data/merge.feather", nthreads=64)

merge = merge[['app', 'ip_os_device_dow_hour']][merge.is_test == False]
merge['counter'] = 1

merge.drop_duplicates(inplace=True)

print(merge.shape)
merge.head()

gc.collect()

merge['counter'] = merge['counter'].astype('bool')

%%time
d = merge.set_index(['ip_os_device_dow_hour', 'app']).unstack(level=-1)

d.head()

applabel = ['app_{0:03d}'.format(i) for i in list(set(merge.app))]
d.columns = applabel
d.head()

d.fillna(False, inplace=True)

print(d.shape)
d.head()

%%time
mat = np.dot(d.astype('bool').astype('uint16').values.T,
             d.astype('bool').astype('uint16').values)

count_app_coincidence_per_user = pd.DataFrame(mat)

del mat, d
gc.collect()

count_app_coincidence_per_user.columns = applabel
count_app_coincidence_per_user.head()

count_app_coincidence_per_user.to_feather('../data/count_app_coincidence_per_user.feather')

del count_app_coincidence_per_user
gc.collect()

ls -lah ../data/*.feather

### pca

import pandas as pd
import numpy as np
from sklearn.decomposition import PCA

d = pd.concat([
    pd.read_feather('../data/agg_by_app.feather'),
    pd.read_feather('../data/uqcount_by_app.feather'),
    pd.read_feather('../data/count_user_by_app_channel.feather'),
    pd.read_feather('../data/count_app_coincidence_per_user.feather')
], axis = 1)

print(d.shape)
d.head()

y = d[['app', 'attrib_by_app']].iloc[:, 1:]
X = d.drop(['app', 'attrib_by_app'], axis=1)

n_components = 20
pca = PCA(n_components = n_components)
res = pca.fit_transform(X.values)

res.shape

pca.explained_variance_ratio_.cumsum()

col = ['app_pca_{0:02d}'.format(i) for i in [i + 1 for i in range(20)]]
res = pd.DataFrame(res, columns=col)
app_pca = pd.concat([y.app, res], axis=1)

app_pca.to_feather('../data/app_pca.feather')

app_pca.head()

### とりあえず回帰してみる

import statsmodels.api as sm

X1 = sm.add_constant(
    app_pca.iloc[:, 1:] / app_pca.iloc[:, 1:].max().max()
    , prepend=False
)

X1.head()

model = sm.GLM(y.attrib_by_app.values,
               X1.values,
               family=sm.families.Poisson(),
               offset = np.log(X.count_by_app.values)
              )

print(y.attrib_by_app.shape)
print(X1.shape)
print(X.count_by_app.shape)

print(model.fit().summary())
